fix: show calendar events in app by syncing Room with backend
Always save to app Room DB and server regardless of device calendar permission; sync server events on calendar screen open; unify bookmark calendar flow through CalendarService.

+++ app/src/main/java/com/wiseyoung/pro/data/model/CalendarModels.kt
@@ -18,7 +18,7 @@ data class CalendarEventRequest(
 )
 
 data class CalendarEventResponse(
-    val eventId: Int,
+    val eventId: Long,
     val userId: String,
     val title: String,
     val eventType: String,
